Relocate encode to a shared library
