+++ stripecardscan/src/main/java/com/stripe/android/stripecardscan/framework/api/Network.kt
@@ -4,8 +4,8 @@ import android.util.Log
 import com.stripe.android.camera.framework.time.Duration
 import com.stripe.android.camera.framework.time.seconds
 import com.stripe.android.stripecardscan.framework.api.StripeNetwork.Companion.RESPONSE_CODE_UNSET
-import com.stripe.android.stripecardscan.framework.util.decodeFromJson
-import com.stripe.android.stripecardscan.framework.util.encodeToXWWWFormUrl
+import com.stripe.android.core.utils.decodeFromJson
+import com.stripe.android.core.utils.encodeToXWWWFormUrl
 import com.stripe.android.stripecardscan.framework.util.retry
 import kotlinx.serialization.KSerializer
 import java.io.File
